requirements.txt
fastapi==0.114.0
uvicorn==0.30.6
openai-whisper==20231117
pydub==0.25.1
python-multipart==0.0.9
